urcu/flip-latch: harden transacted-struct allocation and tagging
The flip-latch engines tag low pointer bits of the records they park in
slots, and document up to 4 free low bits for embedders.  But those structs
were malloc()'d, and __attribute__((aligned(16))) does not bind malloc /
realloc -- they only promise max_align_t (8 on some 32-bit ABIs) -- so the
guarantee was an over-claim.

Allocate the descriptor (lock-free) and the latch array (single-updater)
with posix_memalign(16): we own those allocations, so the 16-byte base is
now portable rather than luck.  There is no aligned realloc, so grow()
allocates a fresh aligned block, copies the records buffered so far, and
frees the old one (safe: the array is unpublished before commit).  The
engine's hard requirement stays bit 0; the 16-byte / 4-bit room is the
opportunistic embedder bonus, so a looser allocator is tolerated and no
alignment assertion is added.

Also enforce the engine's pairwise-distinct-slot precondition with a debug
assertion after the commit-time sort, and tag pointers in the single-updater
list through uintptr_t (not unsigned long), matching the lock-free siblings.

+++ include/urcu/rcu-bidir-list.h
@@ -80,6 +80,7 @@
  */
 
 #include <stdlib.h>
+#include <stdint.h>
 
 #include <urcu/compiler.h>
 #include <urcu/uatomic.h>
@@ -123,7 +124,7 @@ void cds_bidir_list_init(struct cds_bidir_list_head *head)
 static inline
 void *cds_bidir_list_proxy_tag(struct urcu_flip_proxy *proxy)
 {
-	return (void *) ((unsigned long) proxy | CDS_BIDIR_LIST_PROXY_TAG);
+	return (void *) ((uintptr_t) proxy | CDS_BIDIR_LIST_PROXY_TAG);
 }
 
 /*
@@ -134,11 +135,11 @@ void *cds_bidir_list_proxy_tag(struct urcu_flip_proxy *proxy)
 static inline
 struct cds_bidir_list_head *cds_bidir_list_resolve(struct cds_bidir_list_head *ptr)
 {
-	unsigned long v = (unsigned long) ptr;
+	uintptr_t v = (uintptr_t) ptr;
 
 	if (caa_unlikely(v & CDS_BIDIR_LIST_PROXY_TAG)) {
 		struct urcu_flip_proxy *proxy = (struct urcu_flip_proxy *)
-				(v & ~CDS_BIDIR_LIST_PROXY_TAG);
+				(v & ~(uintptr_t) CDS_BIDIR_LIST_PROXY_TAG);
 
 		return (struct cds_bidir_list_head *)
 				urcu_flip_proxy_get(proxy);
